# backend/services/contracts/internal-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Vocabulary Internal Service Contracts
  version: 1.1.0
  description: |
    Source of truth for `/internal/v1/*` contracts used between TypeScript owner-services,
    the public gateway, and Python capability APIs.
servers:
  - url: /internal/v1
paths:
...
  /system/warmup:
    get:
      summary: NLP warmup status snapshot.
      operationId: getInternalWarmup
      responses:
        '200':
          description: Warmup state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarmupStatus'
        '502':
          description: Upstream capability failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /lexicon/categories:
    get:
      summary: Return available lexicon categories.
      operationId: listLexiconCategories
      responses:
        '200':
          description: Available categories.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryList'
    post:
      summary: Create a lexicon category through the owner service.
      operationId: createLexiconCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Category mutation result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryMutationResult'
        '400':
          description: Invalid category request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /lexicon/entries:
    post:
      summary: Insert a single lexicon entry through the owner service.
      operationId: addLexiconEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LexiconEntryCreateRequest'
      responses:
        '200':
          description: Insert result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationMessage'
        '400':
          description: Invalid entry request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /assignments/scan:
    post:
      summary: Synchronous assignment scan.
      operationId: scanAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentScanRequest'
      responses:
        '200':
          description: Scan result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentScanResult'
...
  /assignments/bulk-rescan:
    post:
      summary: Synchronous bulk rescan.
      operationId: bulkRescanAssignments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkRescanRequest'
      responses:
        '200':
          description: Bulk rescan result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkRescanResult'
...
components:
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
...
    WarmupStatus:
      type: object
      required:
        - running
        - ready
        - failed
        - error_message
        - elapsed_sec
      properties:
        running:
          type: boolean
        ready:
          type: boolean
        failed:
          type: boolean
        error_message:
          type: string
        elapsed_sec:
          type:
            - number
            - 'null'
...
    CategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    CategoryList:
      type: object
      required:
        - categories
      properties:
        categories:
          type: array
          items:
            type: string
    CategoryMutationResult:
      type: object
      required:
        - categories
        - message
      properties:
        categories:
          type: array
          items:
            type: string
        message:
          type: string
    MutationMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    LexiconEntryCreateRequest:
      type: object
      required:
        - category
        - value
      properties:
        category:
          type: string
        value:
          type: string
        source:
          type: string
          default: manual
        confidence:
          type: number
          default: 1
...
    AssignmentScanRequest:
      type: object
      required:
        - content_completed
      properties:
        title:
          type: string
          default: ''
        content_original:
          type: string
          default: ''
        content_completed:
          type: string
    AssignmentMatch:
      type: object
      required:
        - entry_id
        - term
        - category
        - source
        - occurrences
      properties:
        entry_id:
          type: integer
        term:
          type: string
        category:
          type: string
        source:
          type: string
        occurrences:
          type: integer
    AssignmentMissingWord:
      type: object
      required:
        - term
        - occurrences
        - example_usage
      properties:
        term:
          type: string
        occurrences:
          type: integer
        example_usage:
          type: string
    AssignmentDiffChunk:
      type: object
      required:
        - operation
        - original_text
        - completed_text
      properties:
        operation:
          type: string
        original_text:
          type: string
        completed_text:
          type: string
    AssignmentScanResult:
      type: object
      required:
        - assignment_id
        - title
        - content_original
        - content_completed
        - word_count
        - known_token_count
        - unknown_token_count
        - lexicon_coverage_percent
        - assignment_status
        - message
        - duration_ms
        - matches
        - missing_words
        - diff_chunks
      properties:
        assignment_id:
          type:
            - integer
            - 'null'
        title:
          type: string
        content_original:
          type: string
        content_completed:
          type: string
        word_count:
          type: integer
        known_token_count:
          type: integer
        unknown_token_count:
          type: integer
        lexicon_coverage_percent:
          type: number
        assignment_status:
          type: string
        message:
          type: string
        duration_ms:
          type: number
        matches:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentMatch'
        missing_words:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentMissingWord'
        diff_chunks:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentDiffChunk'
    BulkRescanRequest:
      type: object
      properties:
        assignment_ids:
          type: array
          items:
            type: integer
      required:
        - assignment_ids
    BulkRescanResult:
      type: object
      required:
        - success_count
        - failed_count
        - message
      properties:
        success_count:
          type: integer
        failed_count:
          type: integer
        message:
          type: string
...